requirements.txt
asttokens==2.4.1
backcall==0.2.0
bcrypt==4.0.1
blinker==1.6.3
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.1.4
debugpy==1.8.0
decorator==5.1.1
dnspython==2.4.2
email-validator==2.1.0.post1
executing==2.0.1
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
google-api-core==2.20.0
google-api-python-client==2.146.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
greenlet==3.0.1
httplib2==0.22.0
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.26.0
ipython==8.12.3
itsdangerous==2.0.1
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.5.0
jupyter_core==5.5.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
oauthlib==3.2.2
packaging==23.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.4.0
platformdirs==3.11.0
prompt-toolkit==3.0.39
proto-plus==1.24.0
protobuf==5.28.2
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
Pygments==2.16.1
PyJWT==2.9.0
pyparsing==3.1.4
pypatterns-575==2.1.2
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
pywin32==306
pyzmq==25.1.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.22
stack-data==0.6.3
tornado==6.3.3
traitlets==5.13.0
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.2.3
wcwidth==0.2.9
Werkzeug==3.0.1
WTForms==3.1.1
zipp==3.17.0
